+++ src/main/java/com/petservice/main/api/database/dto/BnsVL/BnsVLResponse.java
@@ -1,4 +1,4 @@
-package com.petservice.main.api.database.dto;
+package com.petservice.main.api.database.dto.BnsVL;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
